backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.2.1
python-multipart==0.0.6
starlette==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
cryptography==46.0.3

# HTTP & API
httpx==0.25.2
requests==2.31.0
httptools==0.7.1
websockets==15.0.1

# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
lxml==6.0.2
parsel==1.10.0
w3lib==2.3.1

# Caching & Rate Limiting
redis==7.0.1
slowapi==0.1.9
limits==5.6.0

# Monitoring & Logging
watchfiles==1.1.1

# Email
email-validator==2.3.0
dnspython==2.8.0

# Testing
pytest==9.0.0
pytest-cov==4.1.0
pluggy==1.6.0
iniconfig==2.3.0

# Code Quality
Pygments==2.19.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1
click==8.3.0
colorama==0.4.6
PyYAML==6.0.3
